fix: revisions moving primitive types into their components
