+++ packages/ui/components/primitives/PayloadDate/PayloadDate.stories.tsx
@@ -1,7 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
 
-import type { PayloadDateProps as PayloadDateType } from '../types/PayloadDate';
-
+import type { PayloadDateProps as PayloadDateType } from '.';
 import PayloadDate from '.';
 
 const meta: Meta<PayloadDateType> = {
