--- packages/ui/components/primitives/types/PayloadDate.ts
@@ -1,4 +1,0 @@
-export type PayloadDateProps = {
-  input: string;
-  format?: string;
-};
