--- packages/ui/components/primitives/types/PayloadLink.ts
@@ -1,7 +1,0 @@
-import type { LinkProps } from 'next/link';
-
-export type PayloadLinkProps = LinkProps & {
-  text: string;
-  newTab?: boolean;
-  linkType?: string;
-};
