--- packages/ui/components/primitives/types/PayloadText.ts
@@ -1,3 +1,0 @@
-export type PayloadTextProps = {
-  content: string;
-};
