--- packages/ui/components/primitives/types/ResponsivePayloadImage.ts
@@ -1,1 +1,0 @@
-export { type Image as PayloadImageProps } from '@mono/types/payload-types';
